Clean basic folder structure
