+++ tests/examples/models/lameModel.js
@@ -1,0 +1,5 @@
+var LameModel = {class: "Lame",
+  lame: {type: "String"}
+};
+
+module.exports = LameModel;
