+++ tests/examples/models/nameModel.js
@@ -1,0 +1,5 @@
+var NameModel = {class: "Name",
+  name: {type: "String"}
+};
+
+module.exports = NameModel;
