+++ tests/examples/models/userModel.js
@@ -1,0 +1,7 @@
+var UserModel = {class: "User",
+  name: {type: "String"},
+  email: {type: "String"},
+  password: {type: "String"}
+};
+
+module.exports = UserModel;
