+++ tests/unit.js
@@ -1,6 +1,6 @@
 var assert = require('assert');
+var foie = require('../src/foie');
 
-var foie = function() { };
 describe('Foie', function() {
   it('should be a function', function() {
     assert.equal(true, foie instanceof Function);
